+++ R/congress-package.R
@@ -1,4 +1,5 @@
 ## usethis namespace: start
+#' @importFrom lifecycle deprecated
 #' @importFrom rlang .data
 ## usethis namespace: end
 NULL