+++ NEWS.md
@@ -1,3 +1,9 @@
+# congress 0.0.2
+
+* Addresses changes in first November 2022 API (#1).
+  * `cong_communication()` is deprecated in favor of `cong_house_communication()`, 
+  as a new senate endpoint was added. This can be accessed via `cong_senate_communication()`. 
+
 # congress 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.